test_programs/requirements.txt
numpy>=1.24
pillow>=10.0
pyserial>=3.5
